+++ packages/common/src/components/index.ts
@@ -1,4 +1,5 @@
 export { default as TopSheet } from './topSheet/topSheet'
 export { default as HomePageContent } from './homePageContent/homePageContent'
-export { default as PoweredBy } from './poweredBy'
 export { default as SearchAndDiscover } from './searchAndDiscover/searchAndDiscover'
+export { default as TopHeader } from './topHeader/topHeader'
+export { default as SubHeader } from './subHeader/subHeader'
