+++ packages/common/src/components/qrCode/QrCode.module.css
@@ -1,0 +1,3 @@
+.qrCode_tourism canvas {
+  margin: 0 auto;
+}
