+++ packages/common/src/components/settings/settings.module.css
@@ -1,0 +1,22 @@
+.language_toggle_wrapper,
+.language_toggle {
+  position: relative;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+.language_modal {
+  width: max-content;
+  min-width: 184px;
+  position: absolute;
+  right: 0;
+  top: 2rem;
+  background-color: #fff;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  border-radius: 0.5rem;
+  padding-left: 1.5rem;
+  padding-right: 1.5rem;
+  padding-top: 0.75rem;
+  padding-bottom: 0.75rem;
+  z-index: 9999;
+}
