+++ packages/common/src/components/topHeader/constants/index.ts
@@ -1,0 +1,7 @@
+export const appLogoBlackList = ['/signIn', '/signUp']
+export const homeIconBlackList = ['/', '/signIn', '/mobileOtp', '/paymentMode', '/signUp']
+export const menuIconWhiteList = ['/', '/search', '/profile']
+
+export const topHeaderBlackList: string[] = []
+
+export const languageIconWhiteList: string[] = []
