+++ packages/common/src/components/topHeader/topHeader.module.css
@@ -1,0 +1,20 @@
+.top_header {
+  height: 30px;
+  width: 100%;
+  background-color: #efefef;
+  position: fixed;
+  z-index: 9999;
+  /* padding: 0 10rem; */
+  top: 0;
+}
+.top_header_wrapper {
+  display: flex;
+  height: 100%;
+  align-items: center;
+  justify-content: space-between;
+}
+.top_header_modal {
+  display: flex;
+  gap: 8px;
+  padding: 15px 0px;
+}
